% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_model_quality}
\alias{analyze_model_quality}
\title{Analyze Model Quality}
\usage{
analyze_model_quality(pdb)
}
\arguments{
\item{pdb}{pdb object from read.pdb}
}
\value{
Quality assessment results
}
\description{
Analyze Model Quality
}
