% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss-model.R
\name{download_model_file}
\alias{download_model_file}
\title{Download Model Files}
\usage{
download_model_file(coordinates_url, output_file)
}
\arguments{
\item{coordinates_url}{Coordinates file URL}

\item{output_file}{Output file path}
}
\value{
Downloaded file path
}
\description{
Download Model Files
}
