% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss-model.R
\name{get_model_coordinates_urls}
\alias{get_model_coordinates_urls}
\title{Get Model Coordinates URLs}
\usage{
get_model_coordinates_urls(status_info, file_format = "pdb")
}
\arguments{
\item{status_info}{Project status information object}

\item{file_format}{File format ("pdb" or "cif", default: "pdb")}
}
\value{
List of model coordinate URLs
}
\description{
Get Model Coordinates URLs
}
