% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{pdb_info}
\alias{pdb_info}
\title{Protein Structure Information}
\usage{
pdb_info(pdb)
}
\arguments{
\item{pdb}{pdb object from read.pdb}
}
\value{
Structure information
}
\description{
Protein Structure Information
}
