% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_pdb}
\alias{plot_pdb}
\title{Visualize 3D Structure}
\usage{
plot_pdb(pdb)
}
\arguments{
\item{pdb}{pdb object from read.pdb or path to PDB file}
}
\value{
3D visualization object
}
\description{
Visualize 3D Structure
}
