% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_ramachandran}
\alias{plot_ramachandran}
\title{Plot Ramachandran Diagram}
\usage{
plot_ramachandran(pdb)
}
\arguments{
\item{pdb}{pdb object from read.pdb}
}
\value{
Plot object
}
\description{
Plot Ramachandran Diagram
}
