% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_residue_composition}
\alias{plot_residue_composition}
\title{Create Residue Composition Plot}
\usage{
plot_residue_composition(pdb)
}
\arguments{
\item{pdb}{pdb object from read.pdb}
}
\value{
Plot object
}
\description{
Create Residue Composition Plot
}
