% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read.pdb}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bio3d}{\code{\link[bio3d]{read.pdb}}}
}}

