% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss-model.R
\name{run_automodel_workflow}
\alias{run_automodel_workflow}
\title{Complete Automodel Workflow}
\usage{
run_automodel_workflow(
  protein_sequence,
  project_title = "R Automodel",
  output_dir = "./modeling_results",
  wait_interval = 30,
  max_wait_time = 1800
)
}
\arguments{
\item{protein_sequence}{Protein sequence}

\item{project_title}{Project title}

\item{output_dir}{Output directory}

\item{wait_interval}{Check interval}

\item{max_wait_time}{Maximum wait time}
}
\value{
Modeling results
}
\description{
Complete Automodel Workflow
}
