% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss-model.R
\name{run_swiss_model}
\alias{run_swiss_model}
\title{Complete Swiss Model Workflow}
\usage{
run_swiss_model(
  protein_sequence,
  project_title = "R Automodel",
  output_dir = "./modeling_results",
  wait_interval = 30,
  max_wait_time = 1800
)
}
\arguments{
\item{protein_sequence}{Protein sequence}

\item{project_title}{Project title}

\item{output_dir}{Output directory}

\item{wait_interval}{Check interval}

\item{max_wait_time}{Maximum wait time}
}
\value{
Modeling results
}
\description{
Complete Swiss Model Workflow
}
