% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_swissmodel_token}
\alias{set_swissmodel_token}
\title{Set API Token}
\usage{
set_swissmodel_token(token)
}
\arguments{
\item{token}{API token string}
}
\value{
No return value
}
\description{
Set API Token
}
