% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss-model.R
\name{submit_alignment}
\alias{submit_alignment}
\title{Submit Alignment Project}
\usage{
submit_alignment(
  target_sequences,
  template_sequence,
  pdb_id,
  chain_id,
  project_title = "R Alignment Project",
  template_seqres_offset = 0,
  assembly_id = 1
)
}
\arguments{
\item{target_sequences}{Target protein sequence}

\item{template_sequence}{Template protein sequence}

\item{pdb_id}{Template PDB ID}

\item{chain_id}{Chain identifier}

\item{project_title}{Project title}

\item{template_seqres_offset}{Template sequence offset (default: 0)}

\item{assembly_id}{Assembly ID (default: 1)}
}
\value{
Project ID
}
\description{
Submit Alignment Project
}
