% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss-model.R
\name{submit_automodel}
\alias{submit_automodel}
\title{Submit automodel project}
\usage{
submit_automodel(
  target_sequences,
  project_title = "R Modeling Project",
  email = NULL
)
}
\arguments{
\item{target_sequences}{Protein sequence string or vector of strings}

\item{project_title}{Project title}

\item{email}{Email address (optional)}
}
\value{
Project ID
}
\description{
Submit automodel project
}
