% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss-model.R
\name{submit_user_template}
\alias{submit_user_template}
\title{Submit User Template Project}
\usage{
submit_user_template(
  target_sequences,
  template_coordinates_file,
  project_title = "R User Template Project"
)
}
\arguments{
\item{target_sequences}{Target protein sequence}

\item{template_coordinates_file}{Template coordinates file path}

\item{project_title}{Project title}
}
\value{
Project ID
}
\description{
Submit User Template Project
}
