% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissmodel-package.R
\docType{package}
\name{swissmodel-package}
\alias{swissmodel}
\alias{swissmodel-package}
\title{swissmodel: Protein Structure Prediction using Swiss-Model}
\description{
An R interface to the SWISS-MODEL protein structure homology modeling service that provides programmatic access to automated protein structure prediction. This package enables users to predict protein structures from amino acid sequences using the Swiss-Model web service directly within R, while offering basic analysis and visualization capabilities through seamless integration with the 'bio3d' and 'r3dmol' R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/swissmodel}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

}
\keyword{internal}
